requirements_images.txt
# Additional requirements for image processing API
# Install with: pip install -r requirements_images.txt

# Image processing
Pillow>=10.0.0

# Supabase client
supabase>=2.0.0

# Existing dependencies (if not already installed)
fastapi>=0.100.0
uvicorn>=0.20.0
boto3>=1.26.0
pydantic>=2.0.0
python-multipart>=0.0.6
